extract oauth info to module
