+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   match '/*all', to: 'application#cors_preflight_check', via: [:options]
-  resources :integrations, only: %i(show index)
   resources :events, only: %i(index create)
+  resources :integrations, only: %i(index show)
 
   # OAuth Routes
   post '/oauth/request-token', to: 'oauth#request_token'
